+++ src/dcmtkpp/DcmtkAssociation.txx
@@ -9,7 +9,7 @@
 #ifndef _8ac39caa_b7b1_44a8_82fc_e8e3de18b2f8
 #define _8ac39caa_b7b1_44a8_82fc_e8e3de18b2f8
 
-#include "ServiceRole.h"
+#include "DcmtkAssociation.h"
 
 #include <functional>
 #include <utility>
@@ -17,7 +17,6 @@
 #include <dcmtk/config/osconfig.h>
 #include <dcmtk/dcmnet/assoc.h>
 
-#include "dcmtkpp/DcmtkAssociation.h"
 #include "dcmtkpp/message/Message.h"
 #include "dcmtkpp/Network.h"
 
@@ -26,10 +25,10 @@ namespace dcmtkpp
 
 template<typename TMessage>
 TMessage
-ServiceRole
-::_receive(ProgressCallback callback, void* callback_data) const
+DcmtkAssociation
+::receive(ProgressCallback callback, void* callback_data)
 {
-    message::Message const message = this->_receive(callback, callback_data);
+    message::Message const message = this->receive(callback, callback_data);
     return TMessage(message);
 }
