+++ src/dcmtkpp/ServiceRole.h
@@ -27,9 +27,6 @@ namespace dcmtkpp
 class ServiceRole
 {
 public:
-    /// @brief Progress callback, following the semantics of DCMTK.
-    typedef std::function<void(void *, unsigned long)> ProgressCallback;
-    
     /// @brief Create a default Service Role with no network and no association.
     ServiceRole();
 
@@ -56,62 +53,16 @@ public:
     void set_association(DcmtkAssociation * association);
 
 protected:
-    /// @brief Wrapper class for DMCTK progress callbacks.
-    struct ProgressCallbackData
-    {
-        /// @brief Callback function.
-        ProgressCallback callback;
-        
-        /// @brief Callback data.
-        void * data;
-    };
-    
     /// @brief Network used by the ServiceRole.
     Network * _network;
     
     /// @brief Association used by the ServiceRole.
     DcmtkAssociation * _association;
 
-    /// @brief Wrapper from ProgressCallback to DIMSE_ProgressCallback.
-    static void _progress_callback_wrapper(void * data, unsigned long bytes_count);
-    
-    /// @brief Find an accepted presentation context.
-    T_ASC_PresentationContextID _find_presentation_context(
-        std::string const & abstract_syntax) const;
-    
-    /// @brief Send a DIMSE message.
-    void _send(
-        message::Message const & message, std::string const & abstract_syntax,
-        ProgressCallback callback=NULL, void* callback_data=NULL) const;
-    
-    /// @brief Receive a generic DIMSE message.
-    message::Message _receive(
-        ProgressCallback callback=NULL, void* callback_data=NULL) const;
-    
-    /**
-     * @brief Receive a DIMSE message of specific type.
-     *
-     * Throw an exception if the received message is not of the requested type.
-     */
-    template<typename TMessage>
-    TMessage _receive(ProgressCallback callback=NULL, void* callback_data=NULL) const;
-
 private:
-    void _send(
-        DataSet const & obj, T_ASC_PresentationContextID presID,
-        std::string const & transfer_syntax, DUL_DATAPDV pdvType,
-        ProgressCallback callback, void *callbackContext) const;
 
-    std::pair<DataSet, DUL_DATAPDV> _receive_dataset(
-        ProgressCallback callback, void *callbackContext) const;
-    
-    DUL_PDV _read_next_pdv() const;
-
-    void _check_dimse_ready() const;
 };
 
 }
-
-#include "ServiceRole.txx"
 
 #endif // _fa7d372a_dd27_4a1e_9b29_be9d5fbe602a
